Authorization + Notes tag crud (#4)
* Added authorization filters

* Added notes request validation

* Created Tag model

* Added notetag command

* Changed names

* Adedd potential sealed command class

* Added returing tags inside note view

* Notes list now returns tags

* Tags query is now optimized

* Added tags delete from a note

+++ src/main/kotlin/pl/humberd/notesapp/infrastructure/http/note/NoteController.kt
@@ -6,7 +6,11 @@ import org.springframework.web.bind.annotation.*
 import pl.humberd.notesapp.application.command.note.NoteCommandHandler
 import pl.humberd.notesapp.application.command.note.model.NoteCreateCommand
 import pl.humberd.notesapp.application.command.note.model.NoteDeleteCommand
+import pl.humberd.notesapp.application.command.note.model.NoteIsAuthorCommand
 import pl.humberd.notesapp.application.command.note.model.NotePatchCommand
+import pl.humberd.notesapp.application.command.note_tag.NoteTagCommandHandler
+import pl.humberd.notesapp.application.command.note_tag.model.NoteTagCreateCommand
+import pl.humberd.notesapp.application.command.note_tag.model.NoteTagDeleteCommand
 import pl.humberd.notesapp.application.query.note.NoteQueryHandler
 import pl.humberd.notesapp.application.query.note.model.NoteListFilter
 import pl.humberd.notesapp.application.query.note.model.NoteListView
@@ -14,6 +18,8 @@ import pl.humberd.notesapp.application.query.note.model.NoteView
 import pl.humberd.notesapp.infrastructure.common.ResponseBuilder
 import pl.humberd.notesapp.infrastructure.http.note.model.NoteCreateRequest
 import pl.humberd.notesapp.infrastructure.http.note.model.NotePatchRequest
+import pl.humberd.notesapp.infrastructure.http.note.model.NoteTagCreateRequest
+import java.security.Principal
 import javax.validation.Valid
 import kotlin.contracts.ExperimentalContracts
 
@@ -22,13 +28,17 @@ import kotlin.contracts.ExperimentalContracts
 @RequestMapping("/notes")
 class NoteController(
     private val noteCommandHandler: NoteCommandHandler,
-    private val noteQueryHandler: NoteQueryHandler
+    private val noteQueryHandler: NoteQueryHandler,
+    private val noteTagCommandHandler: NoteTagCommandHandler
 ) {
     @PostMapping("")
-    fun create(@RequestBody @Valid body: NoteCreateRequest): ResponseEntity<NoteView> {
-        val note = noteCommandHandler.create(
+    fun create(
+        @RequestBody @Valid body: NoteCreateRequest,
+        principal: Principal
+    ): ResponseEntity<NoteView> {
+        val note = noteCommandHandler.createAndRefresh(
             NoteCreateCommand(
-                authorId = "user-2",
+                authorId = principal.name,
                 title = body.title,
                 url = body.url,
                 content = body.content
@@ -41,21 +51,49 @@ class NoteController(
     }
 
     @GetMapping("/{id}")
-    fun read(@PathVariable("id") id: String): ResponseEntity<NoteView> {
+    fun read(
+        @PathVariable("id") id: String,
+        principal: Principal
+    ): ResponseEntity<NoteView> {
+        noteCommandHandler.ensureIsAuthor(
+            NoteIsAuthorCommand(
+                noteId = id,
+                userId = principal.name
+            )
+        )
+
         return ResponseBuilder.ok(noteQueryHandler.view(id))
     }
 
     @GetMapping("")
-    fun readList(pageable: Pageable): ResponseEntity<NoteListView> {
-        return ResponseBuilder.ok(noteQueryHandler.listView(NoteListFilter(pageable)))
+    fun readList(
+        pageable: Pageable,
+        principal: Principal
+    ): ResponseEntity<NoteListView> {
+        return ResponseBuilder.ok(
+            noteQueryHandler.listView(
+                NoteListFilter(
+                    pageable = pageable,
+                    authorId = principal.name
+                )
+            )
+        )
     }
 
     @PatchMapping("/{id}")
     fun patch(
         @PathVariable("id") id: String,
-        @RequestBody body: NotePatchRequest
+        @RequestBody body: NotePatchRequest,
+        principal: Principal
     ): ResponseEntity<NoteView> {
-        val note = noteCommandHandler.patch(
+        noteCommandHandler.ensureIsAuthor(
+            NoteIsAuthorCommand(
+                noteId = id,
+                userId = principal.name
+            )
+        )
+
+        val note = noteCommandHandler.patchAndRefresh(
             NotePatchCommand(
                 noteId = id,
                 url = body.url,
@@ -71,10 +109,65 @@ class NoteController(
 
     @DeleteMapping("/{id}")
     fun delete(
-        @PathVariable("id") id: String
+        @PathVariable("id") id: String,
+        principal: Principal
     ): ResponseEntity<Unit> {
+        noteCommandHandler.ensureIsAuthor(
+            NoteIsAuthorCommand(
+                noteId = id,
+                userId = principal.name
+            )
+        )
+
         noteCommandHandler.delete(NoteDeleteCommand(id))
 
-        return ResponseBuilder.noContent();
+        return ResponseBuilder.noContent()
+    }
+
+    @PostMapping("/{id}/tags")
+    fun createTag(
+        @PathVariable("id") id: String,
+        @RequestBody body: NoteTagCreateRequest,
+        principal: Principal
+    ): ResponseEntity<NoteView> {
+        noteCommandHandler.ensureIsAuthor(
+            NoteIsAuthorCommand(
+                noteId = id,
+                userId = principal.name
+            )
+        )
+
+        noteTagCommandHandler.createAndRefresh(
+            NoteTagCreateCommand(
+                noteId = id,
+                tagName = body.tagName,
+                userId = principal.name
+            )
+        )
+
+        return ResponseBuilder.ok(noteQueryHandler.view(id))
+    }
+
+    @DeleteMapping("/{noteId}/tags/{tagId}")
+    fun deleteTag(
+        @PathVariable("noteId") noteId: String,
+        @PathVariable("tagId") tagId: String,
+        principal: Principal
+    ): ResponseEntity<Unit> {
+        noteCommandHandler.ensureIsAuthor(
+            NoteIsAuthorCommand(
+                noteId = noteId,
+                userId = principal.name
+            )
+        )
+
+        noteTagCommandHandler.delete(
+            NoteTagDeleteCommand(
+                noteId = noteId,
+                tagId = tagId
+            )
+        )
+
+        return ResponseBuilder.noContent()
     }
 }
